requirements_concise.txt
pip install torch==1.13.0+cu117
pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-1.13.0+cu117.html
pip install torch_geometric
numba==0.55.2
numpy==1.22.4
scikit-learn==1.0
scikit-network==0.31.0
pytorch-lightning==1.5.0
jupyter==1.0.0
jupyter_core==5.3.1
jupyter_server==2.7.3
jupyter==1.0.0
matplotlib==3.4.3
networkx==2.6.3
ninja==1.10.2.3
pandas==1.3.4
pymongo==4.0.1
sacred==0.8.2
scipy==1.10.0
seml==0.3.6
torchtyping==0.1.4
tqdm==4.62.3
typeguard==2.13.0
ripserplusplus @ git+https://github.com/ArGintum/RipserZeros.git@bac8a96f56e9e3ed46202323accbeeee11c4b54c
POT==0.8.2
gudhi==3.5.0